ReservasRestaurante/pyproject.toml
[project]
name = "reservasrestaurante"
version = "0.1.0"
description = ""
authors = [
    {name = "ViniciusMihael",email = "viniciusmihael2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "python-decouple (==3.8)",
]

[tool.poetry]
packages = [{include = "reservasrestaurante", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
